HBASE-18497 Add clock type to proto message for HLC region open/close (revision 4)
- Removed overridden methods in HMaster#get/set/updateClock
- Changed parameter name and exception text in ProtobufUtil#toClockType
- Removed unused import of ClockType in TestMasterNoCluster
- Removed unused ServerManager#sendRegionClose(ServerName, HRegionInfo, Long) that recursively called itself
- Added TODO in RSProcedureDispatcher to update clock with top timestamp of both loops
- Added TODO in RequestConverter for removing/merging masterSystemTime

+++ hbase-server/src/test/java/org/apache/hadoop/hbase/master/TestMasterNoCluster.java
@@ -54,6 +54,7 @@ import org.apache.hadoop.hbase.shaded.protobuf.generated.RegionServerStatusProto
 import org.apache.hadoop.hbase.testclassification.MasterTests;
 import org.apache.hadoop.hbase.testclassification.MediumTests;
 import org.apache.hadoop.hbase.util.FSUtils;
+import org.apache.hadoop.hbase.util.Pair;
 import org.apache.hadoop.hbase.util.Threads;
 import org.apache.hadoop.hbase.zookeeper.MetaTableLocator;
 import org.apache.hadoop.hbase.zookeeper.ZKUtil;
@@ -216,7 +217,7 @@ public class TestMasterNoCluster {
         // Fake a successful close.
         Mockito.doReturn(true).when(spy).
           sendRegionClose((ServerName)Mockito.any(), (HRegionInfo)Mockito.any(),
-            (ServerName)Mockito.any(), (Long)Mockito.any());
+            (ServerName)Mockito.any(), Mockito.any());
         return spy;
       }
